some refactorings of class InputReader; added InputReader::find()

+++ src/InputBuffer.cpp
@@ -25,7 +25,7 @@
 using namespace std;
 
 
-StreamInputBuffer::StreamInputBuffer (istream &is, unsigned bufsize)
+StreamInputBuffer::StreamInputBuffer (istream &is, size_t bufsize)
 	: _is(is), _bufsize(bufsize), _buf1(new UInt8[_bufsize]), _buf2(new UInt8[_bufsize]), _bufptr(_buf1)
 {
 	_size1 = fillBuffer(_buf1);
@@ -63,7 +63,7 @@ int StreamInputBuffer::peek () const {
 
 
 /** Returns the n-th next character without skipping it. */
-int StreamInputBuffer::peek (unsigned n) const {
+int StreamInputBuffer::peek (size_t n) const {
 	if (pos()+n < _size1)
 		return *(_bufptr+n);
 	if (pos()+n < _size1+_size2)
@@ -86,7 +86,7 @@ int StreamInputBuffer::fillBuffer (UInt8 *buf) {
 
 ///////////////////////////////////////////////
 
-SplittedCharInputBuffer::SplittedCharInputBuffer (const char *buf1, unsigned s1, const char *buf2, unsigned s2) {
+SplittedCharInputBuffer::SplittedCharInputBuffer (const char *buf1, size_t s1, const char *buf2, size_t s2) {
 	_buf[0] = buf1;
 	_buf[1] = buf2;
 	_size[0] = buf1 ? s1 : 0;
@@ -111,7 +111,7 @@ int SplittedCharInputBuffer::peek () const {
 }
 
 
-int SplittedCharInputBuffer::peek (unsigned n) const {
+int SplittedCharInputBuffer::peek (size_t n) const {
 	if (n < _size[_index])
 		return _buf[_index][n];
 	n -= _size[_index];
